EVG-16698: Consolidate abstractions around Leafygreen Button
