+++ src/components/ButtonDropdown.tsx
@@ -1,6 +1,6 @@
 import Icon from "@leafygreen-ui/icon";
 import { Menu, MenuItem } from "@leafygreen-ui/menu";
-import { Button } from "components/Button";
+import { LoadingButton } from "components/Buttons";
 
 interface Props {
   disabled?: boolean;
@@ -17,14 +17,14 @@ export const ButtonDropdown: React.VFC<Props> = ({
 }) => (
   <Menu
     trigger={
-      <Button
+      <LoadingButton
         size="small"
         data-cy={dataCy}
         disabled={disabled}
         loading={loading}
       >
         <Icon glyph="Ellipsis" />
-      </Button>
+      </LoadingButton>
     }
     data-cy="card-dropdown"
     adjustOnMutation
