+++ src/components/Buttons/index.tsx
@@ -1,0 +1,2 @@
+export { LoadingButton } from "./LoadingButton";
+export { PlusButton } from "./PlusButton";
