+++ src/components/EditableTagField/TagRow.tsx
@@ -3,8 +3,8 @@ import styled from "@emotion/styled";
 import IconButton from "@leafygreen-ui/icon-button";
 import TextInput from "@leafygreen-ui/text-input";
 import { v4 as uuid } from "uuid";
+import { PlusButton } from "components/Buttons";
 import Icon from "components/Icon";
-import { PlusButton } from "components/Spawn";
 import { size } from "constants/tokens";
 import { ParameterInput, InstanceTag } from "gql/generated/types";
 import { reducer, getInitialState } from "./tagRowReducer";
