+++ src/components/Hosts/HostPopover.tsx
@@ -1,7 +1,7 @@
 import { useState, useRef } from "react";
 import styled from "@emotion/styled";
+import Button from "@leafygreen-ui/button";
 import Popover from "@leafygreen-ui/popover";
-import { Button } from "components/Button";
 import { PopoverContainer } from "components/styles/Popover";
 import { size } from "constants/tokens";
 import { useOnClickOutside } from "hooks";
