+++ src/components/Spawn/Layout.tsx
@@ -1,7 +1,5 @@
-import React from "react";
 import styled from "@emotion/styled";
-import { ExtendableBox } from "@leafygreen-ui/box";
-import Button, { ButtonProps } from "@leafygreen-ui/button";
+import Button from "@leafygreen-ui/button";
 import { Body, H2 } from "@leafygreen-ui/typography";
 import { Table } from "antd";
 import Badge from "components/Badge";
@@ -22,17 +20,6 @@ export const BadgeWrapper = styled.div`
 export const StyledBadge = styled(Badge)`
   margin: 0 ${size.xs};
 `;
-
-export const PlusButton: ExtendableBox<
-  ButtonProps & { ref?: React.Ref<any> },
-  "button"
-> = React.forwardRef(({ leftGlyph, ...rest }: ButtonProps, forwardRef) => (
-  <Button
-    ref={forwardRef}
-    {...{ ...rest, leftGlyph: <Icon glyph="Plus" /> }}
-    as="button"
-  />
-));
 
 const TableContainer = styled.div`
   overflow-x: scroll;
