+++ src/pages/projectSettings/tabs/GeneralTab/Fields/DeactivateStepbackTasksField.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { useMutation } from "@apollo/client";
+import Button from "@leafygreen-ui/button";
 import { Description, Label } from "@leafygreen-ui/typography";
 import { Field } from "@rjsf/core";
-import { Button } from "components/Button";
 import { ConfirmationModal } from "components/ConfirmationModal";
 import ElementWrapper from "components/SpruceForm/ElementWrapper";
 import { useToastContext } from "context/toast";
