+++ src/pages/spawn/spawnHost/SpawnHostButton.tsx
@@ -4,7 +4,7 @@ import styled from "@emotion/styled";
 import Tooltip from "@leafygreen-ui/tooltip";
 import { useLocation } from "react-router-dom";
 import { useSpawnAnalytics } from "analytics";
-import { PlusButton } from "components/Spawn";
+import { PlusButton } from "components/Buttons";
 import { size } from "constants/tokens";
 import { MyHostsQuery, MyHostsQueryVariables } from "gql/generated/types";
 import { GET_MY_HOSTS } from "gql/queries";
